incorporating review comments, and seperating move to seperate class
