+++ tests/unit/test_cli.py
@@ -5,7 +5,7 @@ from databricks.sdk.errors import NotFound
 from databricks.sdk.service import iam
 from databricks.sdk.service.iam import User
 
-from databricks.labs.ucx.cli import migrate_uc_to_uc, repair_run, skip
+from databricks.labs.ucx.cli import move, repair_run, skip
 
 
 def test_skip_no_schema(caplog):
@@ -44,46 +44,44 @@ def test_no_step_in_repair_run(mocker, caplog):
         assert e.args[0] == "You did not specify --step"
 
 
-def test_uc_to_uc_no_ucx(mocker, caplog):
+def test_move_no_ucx(mocker, caplog):
     w = create_autospec(WorkspaceClient)
     w.current_user.me = lambda: iam.User(user_name="foo", groups=[iam.ComplexValue(display="admins")])
     mocker.patch("databricks.labs.ucx.installer.InstallationManager.for_user", return_value=None)
-    migrate_uc_to_uc(w, "", "", "", "", "")
+    move(w, "", "", "", "", "")
     assert [rec.message for rec in caplog.records if "UCX configuration" in rec.message]
 
 
-def test_uc_to_uc_no_catalog(mocker, caplog):
+def test_move_no_catalog(mocker, caplog):
     w = create_autospec(WorkspaceClient)
     w.current_user.me = lambda: iam.User(user_name="foo", groups=[iam.ComplexValue(display="admins")])
     mocker.patch("databricks.labs.ucx.installer.InstallationManager.for_user", return_value=w.current_user)
-    migrate_uc_to_uc(w, "", "", "", "", "")
+    move(w, "", "", "", "", "")
     assert (
         len([rec.message for rec in caplog.records if "Please enter from_catalog and to_catalog" in rec.message]) == 1
     )
 
 
-def test_uc_to_uc_same_schema(mocker, caplog):
+def test_move_same_schema(mocker, caplog):
     w = create_autospec(WorkspaceClient)
     w.current_user.me = lambda: iam.User(user_name="foo", groups=[iam.ComplexValue(display="admins")])
     mocker.patch("databricks.labs.ucx.installer.InstallationManager.for_user", return_value=w.current_user)
-    migrate_uc_to_uc(w, "SrcCat", "SrcS", "*", "SrcCat", "SrcS")
+    move(w, "SrcCat", "SrcS", "*", "SrcCat", "SrcS")
     assert len([rec.message for rec in caplog.records if "please select a different schema" in rec.message]) == 1
 
 
-def test_uc_to_uc_no_schema(mocker, caplog):
+def test_move_no_schema(mocker, caplog):
     w = create_autospec(WorkspaceClient)
     w.current_user.me = lambda: iam.User(user_name="foo", groups=[iam.ComplexValue(display="admins")])
     mocker.patch("databricks.labs.ucx.installer.InstallationManager.for_user", return_value=w.current_user)
-    migrate_uc_to_uc(w, "SrcCat", "", "*", "TgtCat", "")
+    move(w, "SrcCat", "", "*", "TgtCat", "")
     assert len([rec.message for rec in caplog.records if "Please enter from_schema, to_schema" in rec.message]) == 1
 
 
-def test_uc_to_uc(mocker, caplog):
+def test_move(mocker, caplog):
     w = create_autospec(WorkspaceClient)
     w.current_user.me = lambda: iam.User(user_name="foo", groups=[iam.ComplexValue(display="admins")])
     mocker.patch("databricks.labs.ucx.installer.InstallationManager.for_user", return_value=w.current_user)
-    with patch(
-        "databricks.labs.ucx.hive_metastore.table_migrate.TablesMigrate.move_migrated_tables", return_value=None
-    ) as m:
-        migrate_uc_to_uc(w, "SrcC", "SrcS", "*", "TgtC", "ToS")
+    with patch("databricks.labs.ucx.hive_metastore.table_migrate.TableMove.move_tables", return_value=None) as m:
+        move(w, "SrcC", "SrcS", "*", "TgtC", "ToS")
         m.assert_called_once()
